created new deposit records file
